Fix Tic-Tac-Toe.
