+++ 3ds/source/screens/ticTacToeScreen.cpp
@@ -24,6 +24,7 @@
 *         reasonable ways as different from the original version.
 */
 
+#include "structs.hpp"
 #include "screens/screenCommon.hpp"
 #include "screens/ticTacToeScreen.hpp"
 #include "utils/settings.hpp"
@@ -40,6 +41,28 @@
 
 #define COLOR1 C2D_Color32(137, 207, 240, 255) // Color Player 1.
 #define COLOR2 C2D_Color32(3, 192, 60, 255) // Color Player 2.
+
+Structs::TTTPos gameBoardPos[] = {
+	{80,  40, 50, 50}, {135,  40, 50, 50}, {190,  40, 50, 50},
+	{80,  95, 50, 50}, {135,  95, 50, 50}, {190,  95, 50, 50},
+	{80, 150, 50, 50}, {135, 150, 50, 50}, {190, 150, 50, 50},
+};
+
+int gameBoard[] = {
+	0, 0, 0,
+	0, 0, 0,
+	0, 0, 0,
+};
+
+struct tri {
+	int one;
+	int two;
+	int three;
+} winIndexes[] = {
+	{0, 1, 2}, {3, 4, 5}, {6, 7, 8},
+	{0, 3, 6}, {1, 4, 7}, {2, 5, 8},
+	{0, 4, 8}, {2, 4, 6},
+};
 
 bool TicTacToe::touching(touchPosition touch, Structs::TTTPos button) {
 	if (touch.px >= button.x && touch.px <= (button.x + button.w) && touch.py >= button.y && touch.py <= (button.y + button.h))
